+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 # Changelog
 
+## 1.1.0 (\_\_\_)
+
+### New Features
+
+---
+
+- **BREAKING**. **Controls**. Renamed `valueToUrl/valueFromUrl` to `valueToStorage/valueFromStorage`
+- **Combobox controls (Select, SelectMany, Date, Time)**. Added method `.renderPopup()` to have ability to override popup options
+- [SelectControl](src/controls/select.ts). [SelectManyControl](src/controls/selectMany.ts). [RadioControl](src/controls/radio.ts).
+  - Allowed to use mixed types in `$options.items`. Improved JSDoc (added example)
+  - Allowed to prevent closing menu via **ev.preventDefault** => `$options.items = [{value: 1, text: (v, li) => {... li.onclick = (e) => e.preventDefault();... } }]`
+  - Allowed to use complex values with saving to storage (refactored methods `valueToStorage/valueFromStorage`)
+
+### Fixes
+
+---
+
+- **HTML**. Fixed intellisense (html.json was missed in the package)
+- **Styles**. Textbased controls. Added `opacity: 1` for `button[clear]/prefix/postfix` as fix when controls in a flex-row and it resizes on hover
+- **Combobox controls (Select, SelectMany, Date, Time)**. _Memory leak: menu is closed but not removed when click on body outside control_
+
+---
+
+---
+
 ## 1.0.4 (Dec 29, 2023)
 
 ### New Features
@@ -9,7 +34,7 @@
 ---
 
 - [WUPTimeObject](src/objects/timeObject.ts). Add init by date to constructor
-- [DateControl](src/controls/date.ts). Add sync with time (value, min, max)
+- [DateControl](src/controls/date.ts). Add sync with [TimeControl](src/controls/time.ts) (value, min, max)
 
 ---